move norm to core
