+++ src/stream_ml/pytorch/builtin/_norm.py
@@ -2,65 +2,21 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
-from typing import TYPE_CHECKING
-
-from stream_ml.pytorch._base import ModelBase
-from stream_ml.pytorch.builtin._stats.norm import logpdf
-
-if TYPE_CHECKING:
-    from stream_ml.core.data import Data
-    from stream_ml.core.params import Params
-
-    from stream_ml.pytorch.typing import Array
-
 __all__: list[str] = []
 
+from dataclasses import field, make_dataclass
 
-@dataclass(unsafe_hash=True)
-class Normal(ModelBase):
-    r"""1D Gaussian with mixture weight.
+import torch as xp
 
-    :math:`(weight, \mu, \sigma)(\phi1)`
+from stream_ml.core.builtin._norm import Normal as CoreNormal
+from stream_ml.core.typing import ArrayNamespace
 
-    Parameters
-    ----------
-    n_layers : int, optional
-        Number of hidden layers, by default 3.
-    hidden_features : int, optional
-        Number of hidden features, by default 50.
-    sigma_upper_limit : float, optional keyword-only
-        Upper limit on sigma, by default 0.3.
-    fraction_upper_limit : float, optional keyword-only
-        Upper limit on fraction, by default 0.45.s
-    """
+from stream_ml.pytorch._base import ModelBase
+from stream_ml.pytorch.typing import Array, NNModel
 
-    def __post_init__(self) -> None:
-        super().__post_init__()
-
-        # Validate the coord_names
-        if len(self.coord_names) != 1:
-            msg = "Only one coordinate is supported, e.g ('phi2',)"
-            raise ValueError(msg)
-
-    def ln_likelihood(
-        self, mpars: Params[Array], /, data: Data[Array], **kwargs: Array
-    ) -> Array:
-        """Log-likelihood of the distribution.
-
-        Parameters
-        ----------
-        mpars : Params[Array], positional-only
-            Model parameters. Note that these are different from the ML
-            parameters.
-        data : Data[Array]
-            Data (phi1, phi2).
-        **kwargs : Array
-            Additional arguments.
-
-        Returns
-        -------
-        Array
-        """
-        c = self.coord_names[0]
-        return logpdf(data[c], mpars[c, "mu"], self.xp.clip(mpars[c, "sigma"], 1e-10))
+Normal = make_dataclass(
+    "Normal",
+    [("array_namespace", ArrayNamespace[Array], field(default=xp, kw_only=True))],
+    bases=(CoreNormal[Array, NNModel], ModelBase),
+    unsafe_hash=True,
+)
